+++ relay-metrics/src/aggregatorservice.rs
@@ -1,4 +1,3 @@
-use std::iter::FusedIterator;
 use std::time::Duration;
 
 use relay_base_schema::project::ProjectKey;
@@ -11,31 +10,14 @@ use serde::{Deserialize, Serialize};
 use crate::aggregator::{self, AggregatorConfig, ShiftKey};
 use crate::bucket::Bucket;
 use crate::statsd::{MetricCounters, MetricHistograms};
-use crate::{BucketValue, DistributionValue};
 
 /// Interval for the flush cycle of the [`AggregatorService`].
 const FLUSH_INTERVAL: Duration = Duration::from_millis(100);
-
-/// The fraction of [`AggregatorServiceConfig::max_flush_bytes`] at which buckets will be split. A value of
-/// `2` means that all buckets smaller than half of max_flush_bytes will be moved in their entirety,
-/// and buckets larger will be split up.
-const BUCKET_SPLIT_FACTOR: usize = 32;
-
-/// The average size of values when serialized.
-const AVG_VALUE_SIZE: usize = 8;
 
 /// Parameters used by the [`AggregatorService`].
 #[derive(Clone, Debug, Deserialize, Serialize)]
 #[serde(default)]
 pub struct AggregatorServiceConfig {
-    /// The approximate maximum number of bytes submitted within one flush cycle.
-    ///
-    /// This controls how big flushed batches of buckets get, depending on the number of buckets,
-    /// the cumulative length of their keys, and the number of raw values. Since final serialization
-    /// adds some additional overhead, this number is approxmate and some safety margin should be
-    /// left to hard limits.
-    pub max_flush_bytes: usize,
-
     /// Maximum amount of bytes used for metrics aggregation.
     ///
     /// When aggregating metrics, Relay keeps track of how many bytes a metric takes in memory.
@@ -44,12 +26,6 @@ pub struct AggregatorServiceConfig {
     ///
     /// Defaults to `None`, i.e. no limit.
     pub max_total_bucket_bytes: Option<usize>,
-
-    /// The number of logical partitions that can receive flushed buckets.
-    ///
-    /// If set, buckets are partitioned by (bucket key % flush_partitions), and routed
-    /// by setting the header `X-Sentry-Relay-Shard`.
-    pub flush_partitions: Option<u64>,
 
     /// Determines the wall clock time interval for buckets in seconds.
     ///
@@ -119,8 +95,6 @@ pub struct AggregatorServiceConfig {
 impl Default for AggregatorServiceConfig {
     fn default() -> Self {
         Self {
-            max_flush_bytes: 5_000_000, // 5 MB
-            flush_partitions: None,
             max_total_bucket_bytes: None,
             bucket_interval: 10,
             initial_delay: 30,
@@ -218,8 +192,6 @@ pub struct BucketCountInquiry;
 pub struct FlushBuckets {
     /// The project key.
     pub project_key: ProjectKey,
-    /// The logical partition to send this batch to.
-    pub partition_key: Option<u64>,
     /// The buckets to be flushed.
     pub buckets: Vec<Bucket>,
 }
@@ -234,9 +206,7 @@ pub struct AggregatorService {
     aggregator: aggregator::Aggregator,
     state: AggregatorState,
     receiver: Option<Recipient<FlushBuckets, NoResponse>>,
-    max_flush_bytes: usize,
     max_total_bucket_bytes: Option<usize>,
-    flush_partitions: Option<u64>,
 }
 
 impl AggregatorService {
@@ -260,9 +230,7 @@ impl AggregatorService {
         Self {
             receiver,
             state: AggregatorState::Running,
-            max_flush_bytes: config.max_flush_bytes,
             max_total_bucket_bytes: config.max_total_bucket_bytes,
-            flush_partitions: config.flush_partitions,
             aggregator: aggregator::Aggregator::named(name, AggregatorConfig::from(&config)),
         }
     }
@@ -273,30 +241,6 @@ impl AggregatorService {
             .totals_cost_exceeded(self.max_total_bucket_bytes);
 
         sender.send(result);
-    }
-
-    /// Split the provided buckets into batches and process each batch with the given function.
-    ///
-    /// For each batch, log a histogram metric.
-    fn process_batches<F>(&self, buckets: impl IntoIterator<Item = Bucket>, mut process: F)
-    where
-        F: FnMut(Vec<Bucket>),
-    {
-        let capped_batches = CappedBucketIter::new(buckets.into_iter(), self.max_flush_bytes);
-        let num_batches = capped_batches
-            .map(|batch| {
-                relay_statsd::metric!(
-                    histogram(MetricHistograms::BucketsPerBatch) = batch.len() as f64,
-                    aggregator = self.aggregator.name(),
-                );
-                process(batch);
-            })
-            .count();
-
-        relay_statsd::metric!(
-            histogram(MetricHistograms::BatchesPerPartition) = num_batches as f64,
-            aggregator = self.aggregator.name(),
-        );
     }
 
     /// Sends the [`FlushBuckets`] message to the receiver in the fire and forget fashion. It is up
@@ -317,26 +261,18 @@ impl AggregatorService {
         relay_log::trace!("flushing {} projects to receiver", flush_buckets.len());
 
         let mut total_bucket_count = 0u64;
-        for (project_key, project_buckets) in flush_buckets.into_iter() {
-            let bucket_count = project_buckets.len() as u64;
+        for (project_key, buckets) in flush_buckets.into_iter() {
+            let bucket_count = buckets.len() as u64;
             relay_statsd::metric!(
                 histogram(MetricHistograms::BucketsFlushedPerProject) = bucket_count,
                 aggregator = self.aggregator.name(),
             );
             total_bucket_count += bucket_count;
 
-            let partitioned_buckets = self
-                .aggregator
-                .partition_buckets(project_buckets, self.flush_partitions);
-            for (partition_key, buckets) in partitioned_buckets {
-                self.process_batches(buckets, |batch| {
-                    if let Some(ref receiver) = self.receiver {
-                        receiver.send(FlushBuckets {
-                            project_key,
-                            partition_key,
-                            buckets: batch,
-                        });
-                    }
+            if let Some(ref receiver) = self.receiver {
+                receiver.send(FlushBuckets {
+                    project_key,
+                    buckets,
                 });
             }
         }
@@ -414,153 +350,6 @@ impl Drop for AggregatorService {
     }
 }
 
-/// An iterator returning batches of buckets fitting into a size budget.
-///
-/// The size budget is given through `max_flush_bytes`, though this is an approximate number. On
-/// every iteration, this iterator returns a `Vec<Bucket>` which serializes into a buffer of the
-/// specified size. Buckets at the end of each batch may be split to fit into the batch.
-///
-/// Since this uses an approximate function to estimate the size of buckets, the actual serialized
-/// payload may exceed the size. The estimation function is built in a way to guarantee the same
-/// order of magnitude.
-struct CappedBucketIter<T: Iterator<Item = Bucket>> {
-    buckets: T,
-    next_bucket: Option<Bucket>,
-    max_flush_bytes: usize,
-}
-
-impl<T: Iterator<Item = Bucket>> CappedBucketIter<T> {
-    /// Creates a new `CappedBucketIter`.
-    pub fn new(mut buckets: T, max_flush_bytes: usize) -> Self {
-        let next_bucket = buckets.next();
-
-        Self {
-            buckets,
-            next_bucket,
-            max_flush_bytes,
-        }
-    }
-}
-
-impl<T: Iterator<Item = Bucket>> Iterator for CappedBucketIter<T> {
-    type Item = Vec<Bucket>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let mut current_batch = Vec::new();
-        let mut remaining_bytes = self.max_flush_bytes;
-
-        while let Some(bucket) = self.next_bucket.take() {
-            let bucket_size = estimate_size(&bucket);
-            if bucket_size <= remaining_bytes {
-                // the bucket fits
-                remaining_bytes -= bucket_size;
-                current_batch.push(bucket);
-                self.next_bucket = self.buckets.next();
-            } else if bucket_size < self.max_flush_bytes / BUCKET_SPLIT_FACTOR {
-                // the bucket is too small to split, move it entirely
-                self.next_bucket = Some(bucket);
-                break;
-            } else {
-                // the bucket is big enough to split
-                let (left, right) = split_at(bucket, remaining_bytes);
-                if let Some(left) = left {
-                    current_batch.push(left);
-                }
-
-                self.next_bucket = right;
-                break;
-            }
-        }
-
-        if current_batch.is_empty() {
-            // There is still leftover data not returned by the iterator after it has ended.
-            if self.next_bucket.take().is_some() {
-                relay_log::error!("CappedBucketIter swallowed bucket");
-            }
-            None
-        } else {
-            Some(current_batch)
-        }
-    }
-}
-
-impl<T: Iterator<Item = Bucket>> FusedIterator for CappedBucketIter<T> {}
-
-/// Splits this bucket if its estimated serialization size exceeds a threshold.
-///
-/// There are three possible return values:
-///  - `(Some, None)` if the bucket fits entirely into the size budget. There is no split.
-///  - `(None, Some)` if the size budget cannot even hold the bucket name and tags. There is no
-///    split, the entire bucket is moved.
-///  - `(Some, Some)` if the bucket fits partially. Remaining values are moved into a new bucket
-///    with all other information cloned.
-///
-/// This is an approximate function. The bucket is not actually serialized, but rather its
-/// footprint is estimated through the number of data points contained. See
-/// `estimate_size` for more information.
-fn split_at(mut bucket: Bucket, size: usize) -> (Option<Bucket>, Option<Bucket>) {
-    // If there's enough space for the entire bucket, do not perform a split.
-    if size >= estimate_size(&bucket) {
-        return (Some(bucket), None);
-    }
-
-    // If the bucket key can't even fit into the remaining length, move the entire bucket into
-    // the right-hand side.
-    let own_size = estimate_base_size(&bucket);
-    if size < (own_size + AVG_VALUE_SIZE) {
-        // split_at must not be zero
-        return (None, Some(bucket));
-    }
-
-    // Perform a split with the remaining space after adding the key. We assume an average
-    // length of 8 bytes per value and compute the number of items fitting into the left side.
-    let split_at = (size - own_size) / AVG_VALUE_SIZE;
-
-    match bucket.value {
-        BucketValue::Counter(_) => (None, Some(bucket)),
-        BucketValue::Distribution(ref mut distribution) => {
-            let mut org = std::mem::take(distribution);
-
-            let mut new_bucket = bucket.clone();
-            new_bucket.value =
-                BucketValue::Distribution(DistributionValue::from_slice(&org[split_at..]));
-
-            org.truncate(split_at);
-            bucket.value = BucketValue::Distribution(org);
-
-            (Some(bucket), Some(new_bucket))
-        }
-        BucketValue::Set(ref mut set) => {
-            let org = std::mem::take(set);
-            let mut new_bucket = bucket.clone();
-
-            let mut iter = org.into_iter();
-            bucket.value = BucketValue::Set((&mut iter).take(split_at).collect());
-            new_bucket.value = BucketValue::Set(iter.collect());
-
-            (Some(bucket), Some(new_bucket))
-        }
-        BucketValue::Gauge(_) => (None, Some(bucket)),
-    }
-}
-
-/// Estimates the number of bytes needed to serialize the bucket without value.
-///
-/// Note that this does not match the exact size of the serialized payload. Instead, the size is
-/// approximated through tags and a static overhead.
-fn estimate_base_size(bucket: &Bucket) -> usize {
-    50 + bucket.name.len() + aggregator::tags_cost(&bucket.tags)
-}
-
-/// Estimates the number of bytes needed to serialize the bucket.
-///
-/// Note that this does not match the exact size of the serialized payload. Instead, the size is
-/// approximated through the number of contained values, assuming an average size of serialized
-/// values.
-fn estimate_size(bucket: &Bucket) -> usize {
-    estimate_base_size(bucket) + bucket.value.len() * AVG_VALUE_SIZE
-}
-
 /// A message containing a list of [`Bucket`]s to be inserted into the aggregator.
 #[derive(Debug)]
 pub struct MergeBuckets {
@@ -592,7 +381,7 @@ impl MergeBuckets {
 #[cfg(test)]
 mod tests {
 
-    use std::collections::{BTreeMap, BTreeSet};
+    use std::collections::BTreeMap;
     use std::sync::{Arc, RwLock};
 
     use relay_common::time::UnixTimestamp;
@@ -731,224 +520,5 @@ mod tests {
         let bucket_count = aggregator.send(BucketCountInquiry).await.unwrap();
         assert_eq!(bucket_count, 1);
         assert_eq!(receiver.bucket_count(), 0);
-    }
-
-    fn test_config() -> AggregatorServiceConfig {
-        AggregatorServiceConfig {
-            bucket_interval: 1,
-            initial_delay: 0,
-            debounce_delay: 0,
-            max_secs_in_past: 50 * 365 * 24 * 60 * 60,
-            max_secs_in_future: 50 * 365 * 24 * 60 * 60,
-            max_name_length: 200,
-            max_tag_key_length: 200,
-            max_tag_value_length: 200,
-            max_project_key_bucket_bytes: None,
-            max_total_bucket_bytes: None,
-            max_flush_bytes: 50_000_000,
-            ..Default::default()
-        }
-    }
-
-    #[must_use]
-    pub fn run_test_bucket_partitioning(flush_partitions: Option<u64>) -> Vec<String> {
-        let config = AggregatorServiceConfig {
-            max_flush_bytes: 1000,
-            flush_partitions,
-            ..test_config()
-        };
-
-        let bucket1 = Bucket {
-            timestamp: UnixTimestamp::from_secs(999994711),
-            width: 0,
-            name: "c:transactions/foo".to_owned(),
-            value: BucketValue::counter(42.),
-            tags: BTreeMap::new(),
-        };
-
-        let bucket2 = Bucket {
-            timestamp: UnixTimestamp::from_secs(999994711),
-            width: 0,
-            name: "c:transactions/bar".to_owned(),
-            value: BucketValue::counter(43.),
-            tags: BTreeMap::new(),
-        };
-
-        let mut aggregator = AggregatorService::new(config.clone(), None);
-        let project_key = ProjectKey::parse("a94ae32be2584e0bbd7a4cbb95971fed").unwrap();
-        let captures = relay_statsd::with_capturing_test_client(|| {
-            aggregator
-                .aggregator
-                .merge(project_key, bucket1, config.max_total_bucket_bytes)
-                .ok();
-            aggregator
-                .aggregator
-                .merge(project_key, bucket2, config.max_total_bucket_bytes)
-                .ok();
-            aggregator.try_flush();
-        });
-        captures
-            .into_iter()
-            .filter(|x| {
-                [
-                    "metrics.buckets.batches_per_partition",
-                    "metrics.buckets.per_batch",
-                    "metrics.buckets.partition_keys",
-                ]
-                .contains(&x.split_once(':').unwrap().0)
-            })
-            .collect::<Vec<_>>()
-    }
-
-    #[test]
-    fn test_bucket_partitioning_dummy() {
-        let output = run_test_bucket_partitioning(None);
-        insta::assert_debug_snapshot!(output, @r#"
-        [
-            "metrics.buckets.per_batch:2|h|#aggregator:default",
-            "metrics.buckets.batches_per_partition:1|h|#aggregator:default",
-        ]
-        "#);
-    }
-
-    #[test]
-    fn test_bucket_partitioning_128() {
-        let output = run_test_bucket_partitioning(Some(128));
-        // Because buckets are stored in a HashMap, we do not know in what order the buckets will
-        // be processed, so we need to convert them to a set:
-        let (partition_keys, tail) = output.split_at(2);
-        insta::assert_debug_snapshot!(BTreeSet::from_iter(partition_keys), @r#"
-        {
-            "metrics.buckets.partition_keys:59|h|#aggregator:default",
-            "metrics.buckets.partition_keys:62|h|#aggregator:default",
-        }
-        "#);
-
-        insta::assert_debug_snapshot!(tail, @r#"
-        [
-            "metrics.buckets.per_batch:1|h|#aggregator:default",
-            "metrics.buckets.batches_per_partition:1|h|#aggregator:default",
-            "metrics.buckets.per_batch:1|h|#aggregator:default",
-            "metrics.buckets.batches_per_partition:1|h|#aggregator:default",
-        ]
-        "#);
-    }
-
-    #[test]
-    fn test_capped_iter_empty() {
-        let buckets = vec![];
-
-        let mut iter = CappedBucketIter::new(buckets.into_iter(), 200);
-        assert!(iter.next().is_none());
-    }
-
-    #[test]
-    fn test_capped_iter_single() {
-        let json = r#"[
-          {
-            "name": "endpoint.response_time",
-            "unit": "millisecond",
-            "value": [36, 49, 57, 68],
-            "type": "d",
-            "timestamp": 1615889440,
-            "width": 10,
-            "tags": {
-                "route": "user_index"
-            }
-          }
-        ]"#;
-
-        let buckets = serde_json::from_str::<Vec<Bucket>>(json).unwrap();
-
-        let mut iter = CappedBucketIter::new(buckets.into_iter(), 200);
-        let batch = iter.next().unwrap();
-        assert_eq!(batch.len(), 1);
-
-        assert!(iter.next().is_none());
-    }
-
-    #[test]
-    fn test_capped_iter_split() {
-        let json = r#"[
-          {
-            "name": "endpoint.response_time",
-            "unit": "millisecond",
-            "value": [1, 1, 1, 1],
-            "type": "d",
-            "timestamp": 1615889440,
-            "width": 10,
-            "tags": {
-                "route": "user_index"
-            }
-          }
-        ]"#;
-
-        let buckets = serde_json::from_str::<Vec<Bucket>>(json).unwrap();
-
-        // 58 is a magic number obtained by experimentation that happens to split this bucket
-        let mut iter = CappedBucketIter::new(buckets.into_iter(), 108);
-        let batch1 = iter.next().unwrap();
-        assert_eq!(batch1.len(), 1);
-
-        match batch1.first().unwrap().value {
-            BucketValue::Distribution(ref dist) => assert_eq!(dist.len(), 2),
-            _ => unreachable!(),
-        }
-
-        let batch2 = iter.next().unwrap();
-        assert_eq!(batch2.len(), 1);
-
-        match batch2.first().unwrap().value {
-            BucketValue::Distribution(ref dist) => assert_eq!(dist.len(), 2),
-            _ => unreachable!(),
-        }
-
-        assert!(iter.next().is_none());
-    }
-
-    fn test_capped_iter_completeness(max_flush_bytes: usize, expected_elements: usize) {
-        let json = r#"[
-          {
-            "name": "endpoint.response_time",
-            "unit": "millisecond",
-            "value": [1, 1, 1, 1],
-            "type": "d",
-            "timestamp": 1615889440,
-            "width": 10,
-            "tags": {
-                "route": "user_index"
-            }
-          }
-        ]"#;
-
-        let buckets = serde_json::from_str::<Vec<Bucket>>(json).unwrap();
-
-        let mut iter = CappedBucketIter::new(buckets.into_iter(), max_flush_bytes);
-        let batches = iter
-            .by_ref()
-            .take(expected_elements + 1)
-            .collect::<Vec<_>>();
-        assert!(
-            batches.len() <= expected_elements,
-            "Cannot have more buckets than individual values"
-        );
-        let total_elements: usize = batches.into_iter().flatten().map(|x| x.value.len()).sum();
-        assert_eq!(total_elements, expected_elements);
-    }
-
-    #[test]
-    fn test_capped_iter_completeness_0() {
-        test_capped_iter_completeness(0, 0);
-    }
-
-    #[test]
-    fn test_capped_iter_completeness_90() {
-        // This would cause an infinite loop.
-        test_capped_iter_completeness(90, 0);
-    }
-
-    #[test]
-    fn test_capped_iter_completeness_100() {
-        test_capped_iter_completeness(100, 4);
     }
 }